requirements.txt
backports.ssl-match-hostname==3.5.0.1
docker-py==1.10.6
docker-pycreds==0.2.1
ipaddress==1.0.17
requests==2.12.4
six==1.10.0
websocket-client==0.40.0
